requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
PyMySQL==1.1.0
python-dateutil==2.8.2
Flask-Mail==0.10.0
python-dotenv==1.2.2
